requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-surprise
mlflow